Feat: InsightCellView 생성 및 collection, table 셀 분리
